requirements.txt
chardet==5.2.0
charset-normalizer==3.4.3
choreographer==1.0.9
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.59.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
kaleido==0.2.1
kiwisolver==1.4.7
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.9.0
mpld3==0.5.10
numpy==2.0.2
openpyxl==3.1.3
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==5.13.0
pyparsing==3.2.3
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6-WebEngine==6.7.0
PyQt6-WebEngine-Qt6==6.7.3
PyQt6-WebEngineSubwheel-Qt6==6.7.3
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
simplejson==3.20.1
six==1.17.0
tenacity==9.1.2
threadpoolctl==3.6.0
tzdata==2025.2
zipp==3.23.0
